refactor: #8번 피어 리뷰 반영
