+++ src/main/java/com/peppone/dam/board/service/impliments/BoardServiceImpl.java
@@ -1,18 +1,33 @@
 package com.peppone.dam.board.service.impliments;
 
+import static com.peppone.dam.exception.ErrorCode.BOARD_NAME_DUPLICATED;
+import static com.peppone.dam.exception.ErrorCode.BOARD_NOT_FOUND;
+import static com.peppone.dam.exception.ErrorCode.BOARD_URL_DUPLICATED;
+import static com.peppone.dam.exception.ErrorCode.USER_NOT_FOUND;
+import static com.peppone.dam.post.domain.PostType.POST_TYPE_GENERAL;
+import static com.peppone.dam.post.domain.PostType.POST_TYPE_NOTICE;
+import static com.peppone.dam.post.domain.PostType.POST_TYPE_PINNED;
+
 import com.peppone.dam.board.domain.BoardEntity;
 import com.peppone.dam.board.domain.BoardType;
 import com.peppone.dam.board.dto.BoardListDto;
 import com.peppone.dam.board.dto.BoardMakingDto;
 import com.peppone.dam.board.repository.BoardRepository;
 import com.peppone.dam.board.service.BoardService;
-import com.peppone.dam.exception.ErrorCode;
+import com.peppone.dam.exception.CustomException;
+import com.peppone.dam.post.domain.OrderType;
+import com.peppone.dam.post.domain.PostType;
+import com.peppone.dam.post.dto.ReadPostDto;
+import com.peppone.dam.post.repository.PostRepository;
 import com.peppone.dam.response.CommonResponse;
 import com.peppone.dam.response.ResponseService;
 import com.peppone.dam.user.domain.UserEntity;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -22,6 +37,7 @@ public class BoardServiceImpl implements BoardService {
 
   private final BoardRepository boardRepository;
   private final ResponseService responseService;
+  private final PostRepository postRepository;
 
   @Override
   public CommonResponse makeBoard(BoardMakingDto boardMakingDto, UserEntity user) {
@@ -29,15 +45,15 @@ public class BoardServiceImpl implements BoardService {
     boolean boardExistByUrl = boardRepository.existsByUrl(boardMakingDto.getUrl());
 
     if (boardExistByName) {
-      return responseService.ErrorResponse(ErrorCode.BOARD_NAME_DUPLICATED);
+      throw new CustomException(BOARD_NAME_DUPLICATED);
     }
 
     if (boardExistByUrl) {
-      return responseService.ErrorResponse(ErrorCode.BOARD_URL_DUPLICATED);
+      throw new CustomException(BOARD_URL_DUPLICATED);
     }
 
     if (user == null) {
-      return responseService.ErrorResponse(ErrorCode.USER_NOT_FOUND);
+      throw new CustomException(USER_NOT_FOUND);
     }
 
     BoardEntity makeBoard = BoardEntity.builder()
@@ -65,5 +81,65 @@ public class BoardServiceImpl implements BoardService {
         .map(BoardListDto::from).toList();
 
     return responseService.getListResponse(list);
+  }
+
+  @Override
+  public CommonResponse getBoardPostList(String id, long page, long size, OrderType order,
+      boolean orderDirection, Pageable pageable) {
+
+    PageRequest pageRequest = null;
+
+    if (orderDirection) {
+      pageRequest = PageRequest.of((int) page, (int) size,
+          Sort.by(order.toString()).descending());
+    } else {
+      pageRequest = PageRequest.of((int) page, (int) size,
+          Sort.by(order.toString()).ascending());
+    }
+
+    if (!boardRepository.existsByUrl(id)) {
+      return responseService.ErrorResponse(BOARD_NOT_FOUND);
+    }
+
+    BoardEntity board = boardRepository.findByUrl(id);
+
+    List<ReadPostDto> posts = getPostByType(board, pageRequest, POST_TYPE_GENERAL);
+
+    return responseService.getListResponse(posts);
+  }
+
+  @Override
+  public CommonResponse getBoardPinnedPostList(String id) {
+    if (!boardRepository.existsByUrl(id)) {
+      throw new CustomException(BOARD_NOT_FOUND);
+    }
+
+    BoardEntity board = boardRepository.findByUrl(id);
+
+    List<ReadPostDto> posts = getPostByType(board, PageRequest.of(0, 3,
+        Sort.by("id").descending()), POST_TYPE_PINNED);
+
+    return responseService.getListResponse(posts);
+  }
+
+  @Override
+  public CommonResponse getBoardNoticePostList(String id) {
+    if (!boardRepository.existsByUrl(id)) {
+      throw new CustomException(BOARD_NOT_FOUND);
+    }
+
+    BoardEntity board = boardRepository.findByUrl(id);
+
+    List<ReadPostDto> posts = getPostByType(board, PageRequest.of(0, 3,
+        Sort.by("id").descending()), POST_TYPE_NOTICE);
+
+    return responseService.getListResponse(posts);
+  }
+
+  private List<ReadPostDto> getPostByType(BoardEntity board, PageRequest pageRequest,
+      PostType postType) {
+    return postRepository
+        .findAllByBoardIdAndPostTypeAndAccessTrue(board, pageRequest, postType)
+        .stream().map(ReadPostDto::from).toList();
   }
 }
