+++ src/main/java/com/peppone/dam/exception/CustomException.java
@@ -1,0 +1,11 @@
+package com.peppone.dam.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class CustomException extends RuntimeException {
+  private final ErrorCode errorCode;
+
+}
