+++ src/main/java/com/peppone/dam/exception/GlobalExceptionHandler.java
@@ -8,12 +8,17 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @RestControllerAdvice
 @RequiredArgsConstructor
-public class GlobalExceptionHandler extends RuntimeException {
+public class GlobalExceptionHandler {
 
   private final ResponseService responseService;
 
-  @ExceptionHandler(Exception.class)
-  public ErrorResponse Exception(Exception e) {
+  @ExceptionHandler({CustomException.class})
+  protected ErrorResponse handleCustomException(CustomException e) {
+    return responseService.ErrorResponse(e);
+  }
+
+  @ExceptionHandler({Exception.class})
+  protected ErrorResponse handleServerException(Exception e) {
     return responseService.ErrorResponse(e);
   }
