+++ src/main/java/com/peppone/dam/user/controller/UserController.java
@@ -6,6 +6,7 @@ import com.peppone.dam.user.dto.SignInDto;
 import com.peppone.dam.user.service.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -19,21 +20,22 @@ public class UserController {
 
   private final UserService userService;
 
-  @PostMapping("/api/user/sign-in")
+  @PostMapping("/api/user")
   public CommonResponse signIn(@RequestBody @Valid SignInDto signIn) {
     CommonResponse signInResponse = userService.signIn(signIn);
     return signInResponse;
   }
 
-  @GetMapping("/api/user/login")
-  public CommonResponse logIn(@RequestBody @Valid LoginDto login) {
-    CommonResponse loginResponse = userService.logIn(login);
-    return loginResponse;
-  }
-
-  @PatchMapping("/api/user/sign-out")
+  @DeleteMapping("/api/user")
   public CommonResponse signOut(@RequestHeader(name = "Authorization") String token) {
     CommonResponse signOutResponse = userService.signOut(token);
     return signOutResponse;
+  }
+
+
+  @GetMapping("/api/auth")
+  public CommonResponse logIn(@RequestBody @Valid LoginDto login) {
+    CommonResponse loginResponse = userService.logIn(login);
+    return loginResponse;
   }
 }
